+++ cf_xarray/tests/test_units.py
@@ -52,12 +52,12 @@ def test_percent_units():
 @pytest.mark.xfail(reason="not supported by pint, yet: hgrecco/pint#1295")
 def test_udunits_power_syntax():
     """Test that UDUNITS style powers are properly parsed and interpreted."""
-    assert ureg("m2 s-2").units == ureg.m ** 2 / ureg.s ** 2
+    assert ureg("m2 s-2").units == ureg.m**2 / ureg.s**2
 
 
 def test_udunits_power_syntax_parse_units():
     """Test that UDUNITS style powers are properly parsed and interpreted."""
-    assert ureg.parse_units("m2 s-2") == ureg.m ** 2 / ureg.s ** 2
+    assert ureg.parse_units("m2 s-2") == ureg.m**2 / ureg.s**2
 
 
 @pytest.mark.parametrize(
@@ -73,3 +73,11 @@ def test_udunits_format(units, expected):
     u = ureg.parse_units(units)
 
     assert f"{u:cf}" == expected
+
+
+@pytest.mark.parametrize(
+    "alias",
+    [ureg("Celsius"), ureg("degC"), ureg("C"), ureg("deg_C"), ureg("degrees_Celsius")],
+)
+def test_temperature_aliases(alias):
+    assert alias == ureg("celsius")